refactor: 파일 구조 개선을 위해 클래스들을 utils 폴더로 이동
- InputHandler, StringParser, Calculator, OutputHandler를 utils 폴더로 이동
- 관련 import 문 경로 수정
