+++ src/utils/StringParser.js
